+++ src/routes/+page.svelte
@@ -1,6 +1,7 @@
 <script lang="ts">
 	import BodyComposition from '$lib/components/BodyComposition.svelte';
 	import FitnessLevel from '$lib/components/FitnessLevel.svelte';
+	import Forecast from '$lib/components/Forecast.svelte';
 	import { Step, Stepper } from '@skeletonlabs/skeleton';
 </script>
 
@@ -8,6 +9,10 @@
 	<Step>
 		<svelte:fragment slot="header">Result</svelte:fragment>
 		<FitnessLevel />
+	</Step>
+	<Step>
+		<svelte:fragment slot="header">Result</svelte:fragment>
+		<Forecast />
 	</Step>
 	<Step>
 		<svelte:fragment slot="header">Result</svelte:fragment>
