--- src/routes/pratio/+page.svelte
@@ -1,10 +1,0 @@
-<script lang="ts">
-	import GainsCalculator from '$lib/components/gainsCalculator.svelte';
-	import PRatio from '$lib/components/PRatio.svelte';
-
-
-</script>
-
-<PRatio />
-
-<GainsCalculator />
